docker: clarify some of the Dockerfile steps
We don't want to reinstall our gems if the README.md file changes, for
example. Doing the install in a separate dir means that the Docker
filesystem cache stays clean for further steps.

So if I add a new Gem dependency, it will run again but the NPM steps
will still be cached because the /app cache layer is still
clean (since we move ourself to /bundle to do the gems deps).