db: use PostgreSQL
Arguably this covers 90% of usecases out there, and this is just a
template so amend at will.

Also remove mention of running the app locally (i.e without Docker)
since we're going to introduce more services and running them locally
is just a major pain in the neck

People who like to run things locally: just abide and use a safe,
idempotent, contained setup and stop pretending Docker is the issue
and you know better.