docker: use the recommended way of launching container processes
I was trying to understand why Ctrl-C doesn't terminate my Ruby
servers, be it Puma or Webrick. It tries and "gracefully" shut them
down but it always ends up waiting 10 seconds then just killing the
container. Setting up any form of signal trap didn't print anything.

I first came upon a Compose issue[1] which described almost exactly my
issue but turned out to be a false-positive about logging (which is
terminated too early to print proof of logged-signal traps).

But then I finally found a great article from the Yelp people[2] which
gently explains that using

CMD command with some args

spawns the command in a shell, which will have PID 1 whilst the actual
command is forked to 2. This is problematic because the shell process
will trap signals, for reasons I have no idea of.

However

CMD ["command", "with", "some", "args"]

*will* spawn that process, which means it's sitting at the forefront
of the container and will receive signals correctly.

This means a 10-second win on every restart of containers! Bless!

[1]: https://github.com/docker/compose/issues/7135
[2]: https://engineeringblog.yelp.com/2016/01/dumb-init-an-init-for-docker.html#process-behavior-inside-docker-containers